+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- rules are now using dynamic severity which can change at runtime.
 - new class `Attribute` used by `DOMNode`. Attributes now holds the location
   they are created from making DOM attribute rules more precise.
 - new rule `heading-level` for validating sequential heading levels.